pyproject.toml
[project]
name = "clawhub-importer"
version = "0.1.0"
description = "Crawl ClawHub skills and import them into StrawHub with metadata migration"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6",
]

[project.scripts]
clawhub-importer = "clawhub_importer.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawhub_importer"]

[tool.hatch.build.targets.sdist]
include = ["src/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
